Add a validation filter
As NestJS tends to be focussed on JSON APIs, out
of the box, validation returns a 400 error and a
JSON response, which is not helpful for full-stack
apps. This extends the Validation pipe to raise an
error with more context, and allows us to set an
exception filter on a controller by controller level
which will render a template with the errors
available to it.