Add an auto rebase action
We have the repo set up to make sure branches are
up to date before merging. While Github has a
button to update a branch, this adds a merge
commit, which makes the history messy.

This PR adds a Github action that listens for a
comment with `/rebase` in the body and then
automatically rebases against the target branch.

It also automatically rebases if the branch is set to
automerge, meaning if a branch is behind, but has
been approved, you can set it to automerge and
let the robots do the work!