Add ADR for Organisation & Profession data models
This includes a PlantUML diagram to help visualise the models.

This is our first pass at designing our data structures, and will change
over time. We probably won't need to add a new ADR every time this
changes, but it makes sense to keep this decision documented in our
codebase to document our initial assumptions.