Set DB to synchronize locally
This setting:

```
Indicates if database schema should be auto created on every
application launch. Be careful with this option and don't use this in
production - otherwise you can lose production data. This option is
useful during debug and development. As an alternative to it, you can
use CLI and run `schema:sync` command.
```
from: https://typeorm.io/#/connection-options

This essentially means we don't need to write our own migrations at this
stage. We'll need to further investigate how we want to do this for our
production environment.