Allow a user to logout
This adds a form to POST to the inbuilt /logout
route that we get for free with the Okta middleware.
I've styled the button as a link and added the correct
ARIA attributes. I've also added a simple e2e test to
check we can log out. As the application develops, we can add some more tests to check the user's
state, but this is fine as a placeholder.