Log in via Auth0 in e2e tests
This logs into Auth0 via Puppeteer, returning the
cookies and the redirect url. We set these cookies
on the Cypress side and then redirect to the redirect
URL, which then logs us in.

This is all saved using a session, so we only need to
use it once per test run.