Add a guard to check a user is logged in
This adds a `CanActivate` guard (https://docs.nestjs.com/guards)
to check if a user can login. The Open ID middleware
gives us access to an `isAuthenticated()` method
which returns true or false, so we can just check
a) if that method exists on the request and b) if it's
true or false.

We can then use the guard on our controllers like so:

```
@UseGuards(AuthenticationGuard)
```

If the user is logged in, the controller will run, if not
we currently get a JSON 401 response. We'll
probably want to catch this and render something
more friendly.