Get and store the user from the DB on callback
This adds an `afterCallback` function that runs
when `/callback` is visited to fetch the user from the
database and make the object available in the
session. It uses `jwt-decode` to decode the JWT
in the session, get the Auth0 ID from the JSON
blob, and then searches the database for the
relevant user. If no user is present, it raises a
`ForbiddenException`

I've wrapped all this in a class to make it
neater and more testable.