AuthenticationGuard should raise unauthorized
Out of the box, the NestJS guards raise a
`ForbiddenException` when the guard returns false.
At the moment, we're only checking if a user is
logged in, which should return an
`UnauthorizedException`. I've now changed the
guard to specifically raise the error if a user is not
logged in, and made sure the filter catches the
unauthorized error. This has the real life impact of
an unauthenticated user being redirected to the
login path if they try to access an area that requires
login