Add top level info form page for new professions
This is where users will enter the profession's name, regions and select
the relevant industry associated with it.

We're following a pattern of smaller, nested controllers for multi-step
forms, so the TopLevelInformationController will handle this first page
for us.

This currently doesn't submit any information anywhere. We'll need to
store the data in a session to persist it throughout this journey.