Use i18n translations in the application
This moves all the text used in the views in the
application to use translations. This is best
practice because:

a) It allows us to be DRY and if something
changes in one place, we can change it
elsewhere
b) It sets us up to be able to use other
languages (e.g. Welsh) in the future
c) It makes it easier for content designers to
change things in the app without digging
through template files.