Allow users to add top level info for professions
This is a large commit, but it felt necessary to commit all of this at
once as it covers a page transition and form submission.

Top level profession information data is stored in the session, and
played back to the user on the "Check your answers page". In future,
there'll be some more steps in the journey before this page, but it
would be good to keep this simple for the time being.

I've had to rework the way I initially wrote the
`viewTopLevelInformation` function here to use the standard Express
`Response` object rather than Nest's `@Render` decorator to make
re-rendering the page with validation errors as DRY as possible. I
initially tried extending our `ValidationExceptionFilter` but after
struggling with this, I thought it better to at least get something
working for the meantime.

This doesn't currently cover validating that the data in a session is
valid for the step we're curently at, as we do for the User creation
journey. I'd like to add that in a future PR to keep this one as small
and reviewable as possible.