Add optional personalisation args
This adds an added personalisation argument
to the `t` filter that allows us to add
personalisations to our translations. If we have
a translation in one of our json files that looks
like this:

```
// user.json
{
  "greeting": "Hi {user}"
}
```

We can then pass the user's name to the
translation like so:

```
"user. greeting" | t({name: "stuart"})
```

This will output `Hi stuart` to the template