Use `.save`, not `.insert` to create Professions
`.insert` only creates the specified entity, and doesn't run any
callbacks or create any relations, so the Industry ID wasn't being set
as a foreign key here.

Switching to `.save` exposed a flaw in our understanding too, as we
can't create a new `Qualification` here (or any other foreign keys) as
we do in the `Profession` entity's constructor. We've updated the
constructor to set foreign keys as `null` instead when a value isn't
supplied. We'll soon be able to add the real `Qualification`,
`Legislation` etc, so won't need to create a new instance of an object
here anyway - we'll be looking it up.