Allow users to select multiple nations
This was initially an Option Select, which wasn't right, as a profession
could be associated with more than one nation.

The cypress tests for this have been a tricky, for some reason when I
select the single "England" checkbox here, the service throws a CORS
error. I've spent a long time trying to figure out why, but I want to
get this over the line so will hopefully revisit it in future.