Redirect back to the confirm page after changing
This adds a `change=true` param to the edit
urls when making changes. These params are
then sent to the create argument, which then
specify where to redirect. This is not needed
for the roles controller yet, but there's a real
chance we'll be adding more screens, so it
makes sense to be consistent.