List users on an index page
This adds an index page to list all users,
together with a link to view more information
(currently not working!). To get around the
fact that logic in nunjucks templates can get
messy, I've introduced the concept of
presenters, which marshals the user objects
into a format we can use with the the table
macro.