Use Terraform to host app on GPaas
A lot of this setup was copied across from [BEIS
RODA](https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/blob/develop/terraform),
but we've removed things we're not using yet for this project.

This will allow us to deploy both staging and production (and other)
environments to GPaaS by setting the relevant environment variables.