Make all Profession fields nullable
We're planning to move away from the session-based journey for this, as
it will help with editing Professions in future. We'll create a blank
`Profession` on the first step, then add to it on each step of the
journey. We'll mark the Profession as `completed` when we finally submit
on the "Check your answers" page.

I've marked fields as `null` in the constructor here too, including
array values, as checking for `null` (rather than potentially other
values like empty arrays) feels like a consistent way of determining
whether the step has been filled in or not.