Add connect-flash
This will allow us to send flash messages
(for confirmation of updates, deletes etc)
through the stack.

In order to get it to work, I've had to set `esModuleInterop` to true in the tsconfig, so
have has to remove a bunch of
`import * as`s from the main import, but this
seems to work better now.