Stop using sessions for creating a new Profession
We now create a blank Profession at the start of the "Add a profession"
journey and patch that object throughout the process.

This is a fairly large commit as it required a lot of changes to get all
the tests passing in one commit, but some small refactoring, renaming
and test tweaking has snuck in as part of this:
- Rename URLs to be more consistent with the User creation journey, and
  to make it easier for us to re-use them in the "edit" journey later
  on.
- DRY up some of the test setup e.g. stubbing calls to the database.

I've also had to remove the E2E test for viewing a profession as soon as
we create one. This has some potential flakiness to it, as we can't be
sure the Profession would be created with the expected `slug` (locally
for me it was creating it with different one, as though there had been a
collision). This is pretty much covered by
`showing-a-profession.spec.ts`, so I think it's safe to remove for now.