Add a select service page
This adds a controller method, that renders
a form to choose a service if called with no
params, and if called with params, redirects to
the relevant service. If the service is not
recognised, it raises a 404 error.

I've also added the start page and select
service pages to the start of the professions
search e2e tests.