Fix invalid submission of Professions
There was an issue here where the page would hang on submitting without
a required field, due to there not being a `@Render` decorator on the
`edit` function. I've extracted the rendering functionality into its own
function, and tweaked it the logic for what to display.

If there's a blank Profession (i.e. we've just landed on the page)
persist user input even if one or more fields are missing, so they
don't have to re-type their answers.

If there's already a filled-in Profession, e.g. the user is editing
after already completing this step, render those answers on the page,
but if a new answer is submitted and a required field is missing, render
the new values in the form, rather than the older ones on the
Profession.

To make this easier to test, I've made the functions for fetching the
correct value from either the DTO or the Profession object public, this
isn't necessarily the best approach, but we may end up re-using them and
exporting them later for future views.