Add an ENVIRONMENT env var
We have a NODE_ENV var that tells us if the
environment is production, development or
test, but we have no way of telling if our
production environment is staging or prod.
This adds a new env var to our app that tells
us this, allowing us to import different data for
staging and prod environments.