Update professions seeder
Update `ProfessionsSeeder` to assign an organisation to the profession by name.

Also re-order organisations for slightly better demo data, as we will take the first organisation to be the user's profession while we have yet to add an organisation to the user.

Also synchronise test and development seed data where it seems to have become unintentionally out-of-sync