Update Profession with regulated activities form
This follows the usual pattern we've been following of using the DTO
value over any existing Profession value if a new value is submitted.

We're not 100% sure whether the reserved activities should be a
compulsory field here, but we'll leave it as one and see what we gather
in user research.