Add Fishery package
This provides the infrastructure for building Factory objects that build
objects for us that can be used in our tests.

This will be allow us to be consistent with how we set up our object
under test, rather than using a combination of mocks and real objects,
which we've tended to do up until now.