Use factory in Top Level Info controller tests
And, as much as possible move test setup into individual tests, rather
than at the top level, as this groups relevant context together
without having to scroll up and down a file.

This also modifies the factory slightly to allow for us to use
"trait"-style functions - here we do it to add a "blank", just-created
profession.