Add back link to Top Level Information page
This will either take users back to the start of the journey, or to the
check your answers page.

I initially tried using the helper backlink we have in the code, but
this leads to broken journeys when the a previous form page is submitted
with an error. After resubmitting the form successfully but trying to go
back, the page would 404 as it would by trying to visit the `post` URL.