Run deleteUser in the background
This injects a queue into the Auth0Service, and
changes the `deleteUser` method to return
an object with two functions: `performNow`
and `performLater`. The `performLater`
function adds a `deleteUser` job the the
queue, which is then picked up by the
`Auth0Consumer`. This then calls the
`performNow` returned by the `deleteUser`
method.

This allows us to keep all the Auth0-related
code tightly coupled together, rather than
having to add the Auth0 client code to another
Consumer class. The `performLater` and
`performNow` function names also
encapsulate what's going on clearly, as well
as being familiar to Rails devs who are used to
working with ActiveJob.