Add user-friendly error messages
These are a lot nicer to users than the default `x must not be empty`
which often printed out camel case values for `x` which wasn't ideal.

Now we can provide nice, custom error messages. We'll probably want to
get some content eyes on these when we have a content designer in the
team, but this at least demonstrates the pattern for how we'll do this
in future.

This also adds the `tErrors` nunjucks filter to translate an error
object but keep it in the format that the govuk macros expect an
`errorMessage` to be in.