Add presenter for methods to obtain qualifications
This is used to format the data into the right shape for radio buttons
macro args, as well as to pre-check the correct checkbox.

Adding the conditional textarea was a bit fiddly and possibly a bit
hacky here, but it felt a lot cleaner than manually writing out the
all the possible values for the `MethodToObtain` enum. I'd like to have
been able to pass a nunjucks object here, but it's tricky to do in the
views, so have had to opt for HTML.