Move Add a profession start to admin controller
This controller was added in a separate piece of work, but it feels like
a more logical place for these actions to live if they're only going to
be performed by admins.

This also removes the intermediate "start" page for adding a profession,
as it feels unnecessary now we have the "add".

These routes will be placed behind an authentication guard.