Hide create/delete for users without permission
I haven't added any e2e tests for this, because
we don't have enough user types set up to
test this particular use case (i.e. someone who
can manage users, but not create/delete). We
handle this at the controller end, so it feels
like it might be overkill to test with Cypress
too.