Add `admin/professsion` route
This uses the newly-extracted `_profession-details` template to
essentially replay the same template as the public-facing one for now.
We'll be adding a edit and delete buttons to the page very soon.

This is now linked to from the internal admin profession results page,
rather than linking to the public-facing page.