Update govuk-frontend to v4
After adding cypress-axe and running it on some of our e2e test suites,
there were a lot of failures. Particularly to do with things not having
an `aria-label` by default.

After spending a lot of time trying to debug these failures and being
unable to add the necessary attributes via the macros, it turns
out a lot of them are fixed by updating to the next version of
govuk-frontend.