Allow a function to be passed to the interceptor
This allows us to make backlinks
conditional (for example something in the
query string). As we're now dealing with
situations where we may be using
`Response#render` directly, we now need to adjust our appoach a little, and use
the `locals` variable (which will always be
available in the template), rather than
injecting the backUrl directly into the
response.