Test accessibility in `application.spec`
This adds an exclusion to accessibility rules for the phase banner, as
axe was failing tests for the following rule:

"All page content must be contained by landmarks".

This is due to the banner being outside `<header>` and outside `<main>`.
This is a [known
issue](https://github.com/alphagov/govuk-frontend/issues/1604) with the
govuk-frontend component, but as the "region" rule is not a requirement
for WCAG, it should be ok to ignore for the time being on this banner.

I initially tried moving the banner down into `<main>` but this didn't
look good.