Fix / workaround a11y issues on qualification page
There's [a known
issue](https://github.com/alphagov/govuk-frontend/issues/979) with radio
buttons that conditionally-reveal content, due to the `aria-expanded`
attribute being "invalid" on the element. I've decided to ignore that
specific rule on that page here, as we don't want to build our own
custom conditional text box here, and the issue is still under review.

This also fixes an issue with a duplicate `hint` `id` on the page.