Change our exception filter to report to Rollbar
Our HttpExceptionFilter was only catching
HTTP exceptions (as the name, I guess
suggests!), so any errors that weren't
standard HTTP errors were not getting
caught by this filter.

I've tweaked it to now accept all kind of
errors, catch the standard HTTP ones, and
show a friendly error page and log the
error. If we're in production, we also report
the error to Rollbar, as the standard
Express Rollbar middleware wasn't working
as expected