Don't create QualificationPresenter with `null`
If a Profession didn't have a Qualification on it (as is the case with
all the test data), the admin controller would pass `null` into the
`QualificationPresenter`, meaning the check in the template for `if
qualification` would still pass and it would try to call methods on the
`null` qualification. This meant the `view details` links on the admin
side were throwing errors.

I've added a test in to catch this in future, but it shouldn't be a
problem once we get more realistic data in and start building
Qualifications.