Add ProfessionVersion entity
This takes most of the fields from the `Profession` entity, as we'll be
versioning these values.

This makes a couple of changes to the existing Profession data fields,
as it makes sense to start fresh knowing what we know now:

Qualifications and Legislations are now linked to a ProfessionVersion by a
ManyToOne relationship (a profession can have many qualifications and
legislations)

This also adds factory methods for constructing these in tests.