Pass version id along add/edit profession journey
This is a pretty huge commit, but I didn't want to commit any breaking
behaviour, so have done the absolute minimum possible here to set up the
next stage of versioning. Here's what's going on:

- We create a new `ProfessionVersion` when creating a new Profession or
  editing an existing one. In future, this will copy the latest version
  into a new one when editing an existing Profession, but I didn't want
  to overcomplicate things at this stage, and this keeps the journey
  working as expected.
- We pass that `versionId` through the journey in url params, so the
  version can be looked up on every page.
- We don't currently set any values on the `ProfessionVersion`, we
  simply pass it through the journey.

Up next, we'll start populating fields on the ProfessionVersion and stop
accessing them on the Profession.