Lookup versions on profession when editing
Before this change, whenever the profession was updated it would
overwrite with empty professions, as we weren't fetching them on each
page.

This is a temporary step, as once we move fields off the Profession and
onto the ProfessionVersion, we'll only need to look up the
ProfessionVersion itself (which we can get from params), which should
save us some database queries.

This also refactors the ConfirmationController tests to be a bit more in
line with how we're structuring our tests, by setting up context in each
test rather than in the `beforeEach` block.