Copy latest version when editing a profession
This adds a new `ProfessionVersionsController`, which replaces the
existing `edit` and `update` methods that used to live on the
`ProfessionsController`. This is consistent with how we've done things
for the OrganisationVersions too.

We un-set the `id`, `status` (which defaults to `unconfirmed` in the
database), the `created_at` and the `updated_at` fields when saving the
copy.

We also do the same for a Qualification and Legislations here, as we
don't want to update the existing linked Qualification on the old,
still-published `Profession` and show the immediately updated version on
the live page.

This adds `cascade: true` to `Legislations` on the Version to allow us
to save a Legislation when saving a ProfessionVersion.