Store legislations on Version
and read back values from the version itself, rather than the
Profession.

We're currently only storing one Legislation on the Version, but new
designs will allow users to allow an optional extra one or two
Legislations, so for now I'm defaulting to the first element in the
array.

Also play back these answers on the "Check your answers" page, so as not
to break the existing journey (that would otherwise be trying to read
these values from the Profession).

I've left the fields on the Profession entity here, but will remove them
once the full journey is in.