Refactor search controller specs
To have less top-level setup, as it was a bit difficult to figure out
what was going on in each test. I've also introduced slightly more
explicit variable names where relevant.

Although there are a lot more lines in here, it feels a bit easier to
read. I'd like to stop us asserting on the output of
`SearchPresenter.present()` in future, as we already have that tested
elsewhere, and this feels like a lot of complexity to test a controller
for here. However, I didn't want to go down that rabbit hole now, as
stubbing the constructor can be a bit tricky in these tests and we're a
bit tight for time at the moment.