Add an `allDraftOrLive` service method
This fetches all the versions, and adds a
DISTINCT ON query to make sure we only
get one version per organisation. This is
then ordered by the created_at date to
make sure the version we get is the latest
one. We then map the versions to create
on organisation class instance with all the
expected data.