Update a version and an organisation
This updates the `update` action to
fetch an organisation and a version, and
then update the name (if the name has not
already been set), and the
organisationVersion. If the confirm button
is hit, we save both objects and redirect to
the confirmation page.

The next step will be to create a slug for
the organisation and set the version as a
draft when confirming.