Show draft and live professions on the admin index
and display the status of the latest version in the table.

This adds the versions `status` field to the `Profession.withVersion`
function, and to the `Profession` entity itself, so we can treat them as
a single thing.

I've updated the e2e specs to be a bit more specific here. They're not
quite as thorough as the Organisation index tests, which loops through
every seeded Organisation, but I wanted to keep these as fast as
possible here, and not create too much work for us here.

I've also, possibly over-cautiously, escaped the slug here, as we're
outputting direct HTML on the template here.