Add a publish action to the versions service
This accepts the version we want to publish,
and then finds the currently live version. It then
changes its status to live and the currently live
version's status (if it exists) to archived. This
all happens in a transaction, so both status
changes occur at the same time, and also
ensures that if an error occurs, we don't end
up with a situation where we have two live
versions or content disappearing.