Fetch latest live or draft Prof in Org service
Before this change, we'd be fetching all Professions and then filtering
out the Live & Draft ones on each page, then try to call methods on
`null` values, which broke a large chunk of the service.

I initially wrote this by fetching the latest live or draft Profession
in the controllers, but had to update a bunch of Presenters to pass in
these Professions. It didn't quite feel right to update all of these,
when we can update the code where we actually access a Profession on
an Organisation at entity level.

Although we're fetching Live & Draft profession versions for both public
facing and admin-facing code here, we'll filter out the draft versions
before they get displayed to the public.