Fetch latest Profession versions on show page
for both public and admin-facing pages. This also updates the
admin-facing Organisation page to fetch the latest versions of
professions, rather than the top-level profession itself.

We also move admin/professions/show to version controller to be
consistent with how we've done it on the Organisations side.

This also uses the new `findByIdWithProfession` function to display the
version to the admin user, opting for profession & version ids as URL
params over the slug (which we're still using for the public facing
side) to future proof this if we want to be able to display different
versions to users in future.

This also updates URLs throughout the admin side to use this new
endpoint.