Don't attempt to create an existing user in Auth0
Before, when we were editing a user and
confirmed, the app attempted to create our
user in Auth0 again, which showed an error
message (well, attempted to, the app was also
attempting to render a non-existent template).

This changes this to only create the user in
Auth0 if the external identifier is undefined
(i.e. the user hasn't been created in Auth0), as
well as a bit of tidying up.