Remove Profession's `confirmed` field
Now we determine whether a profession has been confirmed via the
presence of its slug, and whether it has a version that has been
published.

This also removes now-unused service methods that relied on that field,
and removes the value from the seeds.

I've also removed the final, half-complete seed in the Profession data,
as it represents an invalid profession that we don't have in the same
format now that we're using versions.