Don't show draft professions on public org page
I previously introduced a bug where I'd fetch all live and draft
versions of a Profession on the public facing organisation, as I thought
we were filtering them elsewhere.

Now `Organisation.withVersions` takes an optional `showDraftProfessions`
parameter that we pass in on the admin-facing side.

I've also added an e2e test to make sure this doesn't slip in again.

This commit removes the invalid draft profession from the seed data, now
that we're using versions, as this draft profession was in an invalid
state.