Workaround for end-to-end session data being cached across runs of `seed:refresh`
As user data is stored in the session on login, we want to make sure that for organisation-level users, we log in anew for each test. Otherwise, session data may reference an organisation that has since been cleared from the database. A more robust solution may be needed if/when we have more tests using organisation-level users, as logging in for each test will slow down E2E test runs