Move Organisation to top level Profession
rather than versioning it.

This means that updating the Regulatory Body on a Profession in the edit
journey will update the profession immediately, and won't go through the
publishing step. This is a known risk, which we will mitigate with
permissions, and/or not allowing users to edit the Organisation at all.

A workaround for this is to add a new profession and copy the existing
values on the new Profession when creating it, so it shouldn't be a huge
issue.

Editing an organisation on a profession is something that will happen
very rarely, and non-"national admin" users shouldn't be allowed to do
this anyway.

The main driver behind this decision was the complex relationship
between ProfessionVersions, Organisations and Professions, that didn't
really make logical sense, but it brought the issues around permissions
here to light.

This commit also removes some duplication from tests that should only be
checking routing, rather than that an entity is updated.