Display user who last changed a ProfessionVersion
on the admin Profession view.

If no user has updated this, we'll display an empty "Changed by". We
initially planned to put "System" here, but we want to display an empty
row in the table, so it's easy for BEIS to see which Organisations have
updated information about their Professions, so it makes sense to be
consistent across both sides here.