Add ProfessionVersions to user
This was missed in a previous piece of work, and it looks like a copy &
paste error in the ProfessionVersion class was classifying this as an
OrganisationVersion.

For some reason, this isn't generating a TypeORM migration. I can't
figure out why, maybe I'm misunderstanding how this works under the
hood.