Hide UK recognition fields when covering the UK
This hides the `ukRecognition` fields if a
profession covers the whole of the UK. It's
involved a bit of a refactor so we can share
the `isUK` and `allNations` functions across
controllers.

It was a bit tricky to hide the fields in the
check your answers page without having to
unpick a bunch of stuff, so I've instead used
the `govuk-visually-hidden` class to hide the
summary list items here.