Display last updated date on admin profession page
This adds a `lastUpdatedDate` field to the `ProfessionPresenter` that
calls `formatDate` so we can use this throughout the app for
consistency.

I've tweaked translations here to nest them more consistently.

Testing this in e2e tests a bit fiddly due to the seed data always
creating a Profession on the day the tests are run, I hope it's not too
bad practice to format today's date like this.