Tidy up Dockerfile
This tweaks the Dockerfile so we can do
caching properly. All of the layers now inherit
from the base layer and we only copy what
we need from the dependency layer,
ensuring that we cache as much as is
humanly possible.

The only things that should not be cached
when running CI are the test dependencies.
This should keep Docker build times down to
a couple of minutes at the most, as long as
the package.json or Dockerfile doesn’t
change.