Deprecate methods to obtain qualification
These will now be inputted via a textarea rather than selecting from a
list of radio buttons.

We can use the existing `otherMethodToObtain` and
`otherCommonPathToObtain` fields for capturing these values, and rename
them now that we've deprecated these fields.

This also removes old fields from the factory that it looks like we
forgot to remove when we previously deprecated these fields.