Use textareas for `routesToObtain` values
this was originally a set of radio buttons, but users found these
confusing and unhelpful and usually chose the conditional "other" one.
Now we just allow them to enter text for each box.

This is a large commit, as it's involved renaming localisation ids,
updating views e.g. the `show` templates, but I didn't want to break any
tests with this change.