Add `summaryList` method to QualificationPresenter
This will replace the nunjucks template's construction of a summaryList
on the `show` pages for both admins and normal users.

Moving the logic here means we can test it, and extend it to
conditionally show fields to members of the public (who won't see blank
fields) or admins (who will see an empty field next to the label), but
that behaviour can come later.