Make Qualification entity fields nullable
There's a chance these won't be present on the data import, so creating
a new Version when editing a Profession would fail at DB-validation
level as it was trying to create a blank Qualification with these fields
set to `null`.

We enforce the mandatory values at code level where necessary, so this
shouldn't be an issue.