Allow users to edit blank Qualifications
When editing a profession with no Qualification set by the data import,
we currently create a blank Qualification, so we're always passing a
`Qualification` (rather than `undefined` or `null`) to the "Check your
answers" page.

This commit tweaks the `QualificationPresenter` to handle missing values
on the `Qualification`, so we can use it whether it has data on it or
not.

We also add a seed Qualification to our test environment here with no
Qualification set, to make sure we can verify this with behaviour our
e2e tests.

This also fixes an out of date test that relied on `industries` being
stored on the top level profession, rather than the `version` that was
brought to light by not putting the old fields on the new test
Profession.