Re-add backlinks
These were accidentally removed when we added an admin header, as the
block overrode the `block` in the `base.njk` in the admin template. I've
extracted the phase banner and backlinks to their own partial here so we
can extend them in both `admin` and `base` templates.

This adds a quick check in the Org "show" e2e spec to click on a back
link, just so this can be caught if it happens again in future for any
reason.