Fix incorrect localisation ids
We moved the localisation id location in a previous commit, but didn't
update it on the public facing page. This moves the search results
localisation id to the `search` key in the translation file, as it's
reused across public and admin pages.

We can pass in HTML as a `count` here so we can customise how it looks
on the public facing page.

We also add a couple of e2e specs to ensure we don't miss this again.