Use the post-validation DTO in controllers
We use the post-validation DTO in our controllers, which will include any changes that occur due to `@Transform` decorators. This involves some update to our controller tests. As before, our controller tests remain somewhat “sociable” in that they use the real `Validator`, as this exercises the decorators on our DTOs