Display profession name when editing
It was unclear to users in this journey exactly what they were doing or
where they were with little context e.g. in case they forget which
Profession they're editing.

We've modified the `captionText` here to pass in the `i18nService` so we
can pass a name to the `translate` function. This means we've had to
pass in the service to each controller used in the `edit` journey.