Get industries from Profession's Versions on Orgs
This wasn't updated when we started storing an industry on a
ProfessionVersion, rather than the Top-level profession.

I think we didn't notice it until now as we haven't updated the seed
data to stop storing this data on a top level Profession.

Now we fetch the latest live/draft version of a Profession and its
Version to read the industries off that.

We no longer need to check that we're fetching a Profession's industries
etc here, as we're doing that as part of the Service method, so I've
removed the tests that check we're throwing an error as they feel a bit
redundant now.