Remove top level profession fields from seeds
These should (hopefully) no longer be used anywhere as they're now on
the "version" entities. If they are, it'll be easy to catch them now.

I've also removed `alternateName` from the top level seeding here, as
even though we're not actually setting it anywhere, it seems like the
sort of thing we'll want to version, so if/when we do add it, we'll put
it on the ProfessionVersion.