Remove no-longer-needed fields from Qualifications
- Qualification level
- Most common route to obtain
- Education duration
- Existence of professional experience required

This is a large commit, but I wanted to commit it all in one to ensure
we didn't break any tests.

We're now not displaying these Qualification fields anywhere e.g. on the
"show" pages or the Check your answers page when adding/editing a
profession.

I've had to update the seeds here, as we can no longer rely on a
qualification level to find the relevant qualification to set on a
profession, so I've gone for using the `routeToObtain` value here.